refactor: Fix conditional rendering of bundle purchase button
This commit fixes the conditional rendering of the bundle purchase button in the ContentsScreen component. Previously, the button was being displayed when the bundle was bought, but it should only be displayed when the bundle is not bought. The logic has been updated to correctly render the button based on the bundle purchase status.