refactor: Update OTP sender to include SMSKEY in console log
This commit updates the OTP sender function in the otpSender.js file to include the SMSKEY in the console log message. This will help with debugging and tracking the SMSKEY being used for sending OTPs. The change is made in the sendOtp function where the console log message is updated to include the SMSKEY variable.