refactor: Update app.json and screens for improved UI and functionality
This commit updates the app.json file to use a new image for the Android adaptive icon. The foregroundImage property now points to "./assets/icons/app.jpg" instead of "./assets/App.png". This change improves the appearance of the app icon on Android devices.

In addition, several screens have been modified to enhance the user experience. The QuizScreen component now has a paddingBottom of 30, providing more space at the bottom of the screen. The HomeScreen component has a paddingTop of 50 and the LoginScreen component has updated default values for the phoneNumber and password fields.

These changes improve the UI and functionality of the app, ensuring a better user experience.