refactor: Update BASE_URL in apis.js to use production URL
This commit updates the BASE_URL in the apis.js file to use the production URL "https://lms-app-1-eoim.onrender.com" instead of the local development URL "http://127.0.0.1:3000". By using the production URL, the application will now make API calls to the correct server in the production environment. This change ensures that the application functions as expected when deployed to the live server.