```text refactor: Update UserCourses component to display course bundle name instead of course name
This commit updates the UserCourses component to display the course bundle name instead of the course name. The code changes include modifying the JSX to use the "bundleName" property instead of the "name" property when rendering the course details. This change ensures that the correct information is displayed to the user.