refactor: Update package.json and uri.tsx
This commit updates the package.json file in the Backend directory to add the "fs" and "https" dependencies. It also updates the uri.tsx file in the KrishnaAcademyAPP/utils directory to change the SERVER_URI to "https://system.krishnaacademy.in". These changes improve the backend functionality and update the server URI to the correct production URL.